Added accidentally ignored file (bin/rubybackup)
bin/ had been automatically added to .gitignore by Github for Windows.
For Ruby apps, bin/ is where the actual command/program that you're
creating goes. (lib/ basically is a collection of gems, modules and
singleton classes, which the script in bin/ uses.)

Anyway, it's been added back in so the thing will actually work now.
Cool.

Also removed in_memtest.rb, which was a temporary file for testing
in-memory zip archives.

Updated README.